moved tests from test_model_caprieval to test_libalign
